mypage타입 중복코드 chore(50%제거)
